support numeric enum property values >127 (beyond Java byte), define initial Script property values
git-svn-id: https://unicode.org/repos/unicodetools/trunk@742 13e8329f-0b23-4da4-9fe8-d0f6fe080806

+++ unicodetools/org/unicode/text/UCD/TestData.java
@@ -74,7 +74,7 @@ public class TestData implements UCD_Types {
                 continue;
             }
             final String norm = Default.nfkc().normalize(i);
-            final byte script = ucd.getScript(i);
+            final short script = ucd.getScript(i);
             final BitSet scripts = ucd.getScripts(norm, normScripts);
             scripts.clear(UCD_Types.COMMON_SCRIPT);
             scripts.clear(UCD_Types.INHERITED_SCRIPT);
